wip: feat/routeset i BP - move OverrideOpHelper to ./lib/collections (need to be verified if it's the right place)
