+++ meteor/client/lib/Components/LabelAndOverrides.tsx
@@ -4,7 +4,10 @@ import { objectPathGet } from '@sofie-automation/corelib/dist/lib'
 import React, { useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 import { ReadonlyDeep } from 'type-fest'
-import { OverrideOpHelperForItemContents, WrappedOverridableItemNormal } from '../../ui/Settings/util/OverrideOpHelper'
+import {
+	OverrideOpHelperForItemContents,
+	WrappedOverridableItemNormal,
+} from '../../../lib/collections/OverrideOpHelper'
 import { DropdownInputOption, findOptionByValue } from './DropdownInput'
 import { hasOpWithPath } from './util'
