+++ meteor/client/lib/forms/SchemaFormTable/ArrayTable.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from 'react-i18next'
 import {
 	WrappedOverridableItemNormal,
 	OverrideOpHelperForItemContents,
-} from '../../../ui/Settings/util/OverrideOpHelper'
+} from '../../../../lib/collections/OverrideOpHelper'
 import { useToggleExpandHelper } from '../../../ui/util/useToggleExpandHelper'
 import { doModalDialog } from '../../ModalDialog'
 import {
