+++ meteor/client/lib/forms/SchemaFormTable/ArrayTableOpHelper.tsx
@@ -1,5 +1,5 @@
 import { clone, objectPathSet } from '@sofie-automation/corelib/dist/lib'
-import { OverrideOpHelperForItemContents } from '../../../ui/Settings/util/OverrideOpHelper'
+import { OverrideOpHelperForItemContents } from '../../../../lib/collections/OverrideOpHelper'
 
 /**
  * The OverrideOp system does not support Arrays currently.
