+++ meteor/client/lib/forms/SchemaFormTable/ArrayTableRow.tsx
@@ -1,5 +1,5 @@
 import React, { useMemo } from 'react'
-import { WrappedOverridableItemNormal } from '../../../ui/Settings/util/OverrideOpHelper'
+import { WrappedOverridableItemNormal } from '../../../../lib/collections/OverrideOpHelper'
 import { SchemaFormTableEditRow } from './TableEditRow'
 import { SchemaTableSummaryRow } from '../SchemaTableSummaryRow'
 import { literal } from '@sofie-automation/shared-lib/dist/lib/lib'
