+++ meteor/client/lib/forms/SchemaFormTable/ObjectTable.tsx
@@ -7,7 +7,7 @@ import {
 	WrappedOverridableItemNormal,
 	OverrideOpHelperForItemContents,
 	getAllCurrentAndDeletedItemsFromOverrides,
-} from '../../../ui/Settings/util/OverrideOpHelper'
+} from '../../../../lib/collections/OverrideOpHelper'
 import { useToggleExpandHelper } from '../../../ui/util/useToggleExpandHelper'
 import { doModalDialog } from '../../ModalDialog'
 import {
