+++ meteor/client/lib/forms/SchemaFormTable/TableEditRow.tsx
@@ -2,7 +2,7 @@ import { faCheck } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import classNames from 'classnames'
 import React, { useCallback } from 'react'
-import { OverrideOpHelperForItemContents } from '../../../ui/Settings/util/OverrideOpHelper'
+import { OverrideOpHelperForItemContents } from '../../../../lib/collections/OverrideOpHelper'
 import { SchemaFormSofieEnumDefinition } from '../schemaFormUtil'
 import { SchemaFormWithOverrides } from '../SchemaFormWithOverrides'
 import { JSONSchema } from '@sofie-automation/shared-lib/dist/lib/JSONSchemaTypes'
