+++ meteor/server/publications/pieceContentStatusUI/checkPieceContentStatus.ts
@@ -42,6 +42,7 @@ import {
 	PackageInfoLight,
 	PieceDependencies,
 } from './common'
+import { getAllCurrentAndDeletedItemsFromOverrides } from '../../../lib/collections/OverrideOpHelper'
 
 interface ScanInfoForPackages {
 	[packageId: string]: ScanInfoForPackage
@@ -505,7 +506,7 @@ async function checkPieceContentExpectedPackageStatus(
 	let progress: number | undefined
 
 	if (piece.expectedPackages && piece.expectedPackages.length) {
-		const routes = getActiveRoutes(studio.routeSets)
+		const routes = getActiveRoutes(getAllCurrentAndDeletedItemsFromOverrides(studio.routeSets, null))
 
 		for (const expectedPackage of piece.expectedPackages) {
 			// Route the mappings
